Common

SomeGraphs.Common — Module

Common data types and utilities for defining specific graph types.

SomeGraphs.Common.Graph — Type

The type of a figure we can display. This is a combination of some AbstractGraphData and AbstractGraphConfiguration . Accessing the .figure property of the graph will return it as a PlotlyFigure , which can be displayed in interactive environments (Julia REPL and/or Jupyter notebooks).

Note

You should call save_graph to save the graph to a file, instead of calling savefig on the .figure , because the latter does't obey the graph size, because Plotly.

SomeGraphs.Common.AbstractGraphConfiguration — Type

A configuration of a [ Graph ] specifies how to display the data while being (as much as possible) independent of the data itself. That is, it should be possible to apply the same configuration to multiple sets of data to generate multiple similar graphs. In some cases (e.g., colors) you can specify a default in the configuration and override it for specific entities in the data.

SomeGraphs.Common.AbstractGraphData — Type

Some data of a [ Graph ] specifies what to display the data while being (as much as possible) independent of how to display it. That is, it should be possible to apply multiple sets of data to the same configuration to to generate multiple similar graphs. In some cases (e.g., colors) you can specify a default in the configuration and override it for specific entities in the data.

SomeGraphs.Common.PlotlyFigure — Type

The type of a rendered graph which Julia knows how to display.

A plotly figure contains everything needed to display an interactive graph (or generate a static one on disk). It can also be converted to a JSON string for handing it over to a different programming language (e.g., to be used to display the interactive graph in a Python Jupyter notebook, given an appropriate wrapper code).

SomeGraphs.Common.save_graph — Function

save_graph(graph::Graph, output_file::AbstractString)::Nothing

Save the graph to a file. Unlike the Plotly savefig function, this function will actually obey the width and height parameters specified in the graph's configuration. The format is deduced from the suffix of the file name.

SomeGraphs.Common.FigureConfiguration — Type

@kwdef mutable struct FigureConfiguration <: Validated
    margins::MarginsConfiguration = MarginsConfiguration()
    width::Maybe{Int} = nothing
    height::Maybe{Int} = nothing
    template::AbstractString = "simple_white"
    show_grid::Bool = true
    grid_color::AbstractString = "lightgrey"
    show_ticks::Bool = true
end

Generic configuration that applies to the whole figure. Each complete AbstractGraphConfiguration contains a figure field of this type.

The optional width and height are in pixels, that is, 1/96 of an inch. The margins are specified in the same units.

By default, show_grid and show_ticks are set.

The default template is "simple white" which is the cleanest. The `show grid and show ticks can be used to disable the grid and/or ticks for an even cleaner (but less informative) look, and the grid color` can also be changed.

SomeGraphs.Common.MarginsConfiguration — Type

@kwdef mutable struct MarginsConfiguration <: Validated
    left::Int = 50
    bottom::Int = 50
    right::Int = 50
    top::Int = 50
end

Configure the margins of the graph. Sizes are in pixels (1/96th of an inch). Plotly is uncapable of automatically adjusting the margins to adapt to tick labels, so you may have to tweak this manually to avoid clipping or reduce wasted white space. In the 21st century. Sigh.

SomeGraphs.Common.AxisConfiguration — Type

@kwdef mutable struct AxisConfiguration <: Validated
    minimum::Maybe{Real} = nothing
    maximum::Maybe{Real} = nothing
    log_scale::Maybe{LogScale} = nothing
    log_regularization::Real = 0
    percent::Bool = false
end

Generic configuration for a graph axis. Everything is optional; by default, the minimum and maximum are computed automatically from the data.

If log_scale is specified, then the log_regularization is added to the coordinate to avoid zero values, and the axis is shown according to the LogScale . Otherwise, log_regularization must be 0.

If percent is set, then the values are multiplied by 100 and a % suffix is added to the tick labels.

The minimum/maximum, data values, color palette values etc. are all in the original scale. That is, you should be able to control log scale and/or percent scaling without changing anything else.

SomeGraphs.Common.LineConfiguration — Type

@kwdef mutable struct LineConfiguration <: Validated
    width::Real = 1
    style::Maybe{LineStyle} = SolidLine
    is_filled::Bool = false
    color::Maybe{AbstractString} = nothing
end

Configure a line in a graph.

If style is nothing no line is drawn. If is_filled then the region defined by the line (below it in a lines graph) is filled. A line with no style and no filled is not drawn but can still be used to define a region (e.g., for a BandsConfiguration ).

By default, the color is chosen automatically.

SomeGraphs.Common.LogScale — Type

Supported log scales (when log scaling is enabled):

Log10Scale converts values to their log (base 10).

Log2Scale converts values to their log (base 2).

SomeGraphs.Common.LineStyle — Type

Styles of drawing a line

SolidLine draws a solid line (the default).

DashedLine draws a dashed line.

SomeGraphs.Common.ValuesOrientation — Type

The orientation of the values axis in a distribution(s) or bars graph:

HorizontalValues - The values are the X axis

VerticalValues - The values are the Y axis (the default).

SomeGraphs.Common.BandsConfiguration — Type

@kwdef mutable struct BandsConfiguration <: Validated
    low::BandConfiguration = BandConfiguration(is_dashed = true)
    middle::BandConfiguration = BandConfiguration()
    high::BandConfiguration = BandConfiguration(is_dashed = true)
end

Configure the partition of the graph up to three band regions. The low and high bands are for the "outer" regions (so their lines are at their border, dashed by default) and the middle band is for the "inner" region between them (so its line is inside it, solid by default).

SomeGraphs.Common.BandConfiguration — Type

@kwdef mutable struct BandConfiguration <: Validated
    offset::Maybe{Real} = nothing
    line::LineConfiguration = LineConfiguration()
end

Configure a region of the graph defined by some band of values.

The offset specifies the band's defining line position. We allow up to three bands in a complete BandsConfiguration . A band exists only if its offset is specified, in which case the line specifies how to render its defining line. The low and high bands are defined the region below and above their defining line's offset . If both are defined, the middle band offset defines the center line of the band; the band can still be filled even if this offset is not specified.

SomeGraphs.Common.BandsData — Type

@kwdef mutable struct BandsData
    low_offset::Maybe{Real} = nothing
    middle_offset::Maybe{Real} = nothing
    high_offset::Maybe{Real} = nothing
end

Override or specify the offsets for tha bands in the data instead of the configuration. Similarly to colors, whether the offset is a data or a configuration parameters depends on the specific graph.

SomeGraphs.Common.ColorsConfiguration — Type

@kwdef mutable struct ColorsConfiguration <: Validated
    show_legend::Bool = false
    color_axis::AxisConfiguration = AxisConfiguration()
    reverse::Bool = false
    colors_palette::Maybe{Union{AbstractString, ContinuousColors, CategoricalColors}} = nothing
end

Configure how to color some values. The colors_palette is applied; this can be:

The name of a standard Plotly palette (see NAMED_COLOR_PALETTES );

A vector of (value, color) tuples for a continuous (numeric value) scale;

Or a vector of (value, color) for a categorical (string value) scale.

If show_legend is set, then the colors will be shown (in the legend or as a color bar, as appropriate).

The color_axis can be used to tweak (continuous) colors. Specifically, the values to be converted to color are modified according to the log_scale and/or percent flags. This is also done to the values specified in an explicit continuous colors pallete.

In Plotly's API, a color pallete is always specified to cover the values between 0 and 1, and it linearly maps some range of colors to this unit range. You can control the mapping of the values to be colored by using color_axis to modify the values ( log_scale , percent ) and also specify the specific range of color values to map to the unit range ( minimum , maximum ).

When specifying an explict continuous colors pallete, we must map its lowest value to 0 and the highest value to 1 (and the values must be in non-decreasing order). In this case, there's no point in specifying minimum and maximum in the color_axis , as these are already encoded in the pallete itself.

SomeGraphs.Common.ContinuousColors — Type

A continuous colors palette, mapping numeric values to colors. We also allow specifying tuples instead of pairs to make it easy to invoke the API from other languages such as Python which do not have the concept of a Pair .

SomeGraphs.Common.CategoricalColors — Type

A categorical colors palette, mapping string values to colors. We also allow specifying tuples instead of pairs to make it easy to invoke the API from other languages such as Python which do not have the concept of a Pair .

SomeGraphs.Common.NAMED_COLOR_PALETTES — Constant

Builtin color palattes from Plotly , both linear: Blackbody , Bluered , Blues , Cividis , Earth , Electric , Greens , Greys , Hot , Jet , Picnic , Portland , Rainbow , RdBu , Reds , Viridis , YlGnBu , YlOrRd and cyclical: Twilight , IceFire , Edge , Phase , HSV , mrybm , mygbm .

The _r (reversed) variants are not included as explicit entries in the dictionary; they are computed on-the-fly if used.

Note

You would think we could just give the builtin color palette names to plotly, but it turns out that "builtin" in Python plotly doesn't mean "builtin" in JavaScript plotly, because "reasons". We therefore have to copy their definition here. An upside of having this dictionary is that you are free to insert additional named palettes into and gain the convenience of refering to them by name (e.g., for coloring heatmap annotations).

Index

SomeGraphs.Common

SomeGraphs.Common.NAMED_COLOR_PALETTES

SomeGraphs.Common.AbstractGraphConfiguration

SomeGraphs.Common.AbstractGraphData

SomeGraphs.Common.AxisConfiguration

SomeGraphs.Common.BandConfiguration

SomeGraphs.Common.BandsConfiguration

SomeGraphs.Common.BandsData

SomeGraphs.Common.CategoricalColors

SomeGraphs.Common.ColorsConfiguration

SomeGraphs.Common.ContinuousColors

SomeGraphs.Common.FigureConfiguration

SomeGraphs.Common.Graph

SomeGraphs.Common.LineConfiguration

SomeGraphs.Common.LineStyle

SomeGraphs.Common.LogScale

SomeGraphs.Common.MarginsConfiguration

SomeGraphs.Common.PlotlyFigure

SomeGraphs.Common.ValuesOrientation

SomeGraphs.Common.save_graph