Common

SomeGraphs.Common — Module

Common data types and utilities for defining specific graph types.

SomeGraphs.Common.Graph — Type

The type of a figure we can display. This is a combination of some AbstractGraphData and AbstractGraphConfiguration . Accessing the .figure property of the graph will return it as a PlotlyFigure , which can be displayed in interactive environments (Julia REPL and/or Jupyter notebooks).

Note

You should call save_graph to save the graph to a file, instead of calling savefig on the .figure , because the latter does't obey the graph size, because Plotly.

SomeGraphs.Common.AbstractGraphConfiguration — Type

A configuration of a [ Graph ] specifies how to display the data while being (as much as possible) independent of the data itself. That is, it should be possible to apply the same configuration to multiple sets of data to generate multiple similar graphs. In some cases (e.g., colors) you can specify a default in the configuration and override it for specific entities in the data.

SomeGraphs.Common.AbstractGraphData — Type

Some data of a [ Graph ] specifies what to display the data while being (as much as possible) independent of how to display it. That is, it should be possible to apply multiple sets of data to the same configuration to to generate multiple similar graphs. In some cases (e.g., colors) you can specify a default in the configuration and override it for specific entities in the data.

SomeGraphs.Common.PlotlyFigure — Type

The type of a rendered graph which Julia knows how to display.

A plotly figure contains everything needed to display an interactive graph (or generate a static one on disk). It can also be converted to a JSON string for handing it over to a different programming language (e.g., to be used to display the interactive graph in a Python Jupyter notebook, given an appropriate wrapper code).

SomeGraphs.Common.save_graph — Function

save_graph(graph::Graph, output_file::AbstractString)::Nothing

Save the graph to a file. Unlike the Plotly savefig function, this function will actually obey the width and height parameters specified in the graph's configuration. The format is deduced from the suffix of the file name.

SomeGraphs.Common.FigureConfiguration — Type

@kwdef mutable struct FigureConfiguration <: Validated
    margins::MarginsConfiguration = MarginsConfiguration()
    width::Maybe{Int} = nothing
    height::Maybe{Int} = nothing
    show_grid::Bool = true
    grid_color::AbstractString = "lightgrey"
    background_color::AbstractString = "white"
    paper_color::AbstractString = "white"
end

Generic configuration that applies to the whole figure. Each complete AbstractGraphConfiguration contains a figure field of this type.

The optional width and height are in pixels, that is, 1/96 of an inch. The margins are specified in the same units.

By default, show_grid is set. It can be used to disable the grid for an even cleaner (but less informative) look. You can also manually change the grid_color , background_color and paper_color .

Examples:

Default (serves as a baseline to compare with when modifying options):

using SomeGraphs
graph = distribution_graph(; distribution_values = [0, 0, 1, 1, 1, 3])
using PlotlyDocumenter
to_documenter(graph.figure)

Reduced size:

using SomeGraphs
graph = distribution_graph(; distribution_values = [0, 0, 1, 1, 1, 3])
graph.configuration.figure.width = 300
graph.configuration.figure.height = 300
using PlotlyDocumenter
to_documenter(graph.figure)

Disable grid:

using SomeGraphs
graph = distribution_graph(; distribution_values = [0, 0, 1, 1, 1, 3])
graph.configuration.figure.show_grid = false
using PlotlyDocumenter
to_documenter(graph.figure)

Change colors:

using SomeGraphs
graph = distribution_graph(; distribution_values = [0, 0, 1, 1, 1, 3])
graph.configuration.figure.grid_color = "black"
graph.configuration.figure.background_color = "lightyellow"
graph.configuration.figure.paper_color = "lightgrey"
using PlotlyDocumenter
to_documenter(graph.figure)

SomeGraphs.Common.MarginsConfiguration — Type

@kwdef mutable struct MarginsConfiguration <: Validated
    left::Int = 50
    bottom::Int = 50
    right::Int = 50
    top::Int = 50
end

Configure the margins of the graph. Sizes are in pixels (1/96th of an inch). Plotly is uncapable of automatically adjusting the margins to adapt to tick labels, so you may have to tweak this manually to avoid clipping or reduce wasted white space. In the 21st century. Sigh.

Examples:

Increased margins:

using SomeGraphs
graph = distribution_graph(; distribution_values = [0, 0, 1, 1, 1, 3])
graph.configuration.figure.margins.left = 100
graph.configuration.figure.margins.right = 200
graph.configuration.figure.margins.top = 150
graph.configuration.figure.margins.bottom = 250
using PlotlyDocumenter
to_documenter(graph.figure)

SomeGraphs.Common.AxisConfiguration — Type

@kwdef mutable struct AxisConfiguration <: Validated
    minimum::Maybe{Real} = nothing
    maximum::Maybe{Real} = nothing
    log_scale::Maybe{LogScale} = nothing
    log_regularization::Real = 0
    percent::Bool = false
end

Generic configuration for a graph axis. Everything is optional; by default, the minimum and maximum are computed automatically from the data.

If log_scale is specified, then the log_regularization is added to the coordinate to avoid zero values, and the axis is shown according to the LogScale . Otherwise, log_regularization must be 0.

If percent is set, then the values are multiplied by 100 and a % suffix is added to the tick labels.

The minimum/maximum, data values, color palette values etc. are all in the original scale. That is, you should be able to control log scale and/or percent scaling without changing anything else.

Examples:

Override range:

using SomeGraphs
graph = distribution_graph(; distribution_values = [0, 0, 1, 1, 1, 3])
graph.configuration.value_axis.minimum = 1
graph.configuration.value_axis.maximum = 4
using PlotlyDocumenter
to_documenter(graph.figure)

Percent:

using SomeGraphs
graph = distribution_graph(; distribution_values = [0, 0, 1, 1, 1, 3])
graph.configuration.value_axis.percent = 1
using PlotlyDocumenter
to_documenter(graph.figure)

SomeGraphs.Common.LogScale — Type

Supported log scales (when log scaling is enabled):

Log10Scale converts values to their log (base 10).

Log2Scale converts values to their log (base 2).

Examples:

Log 2:

using SomeGraphs
graph = distribution_graph(; distribution_values = [0, 0, 1, 1, 1, 3])
graph.configuration.value_axis.log_scale = Log2Scale
graph.configuration.value_axis.log_regularization = 1
using PlotlyDocumenter
to_documenter(graph.figure)

Log 10:

using SomeGraphs
graph = distribution_graph(; distribution_values = [0, 0, 1, 1, 1, 3])
graph.configuration.value_axis.log_scale = Log10Scale
graph.configuration.value_axis.log_regularization = 1e-5
using PlotlyDocumenter
to_documenter(graph.figure)

SomeGraphs.Common.SizeConfiguration — Type

@kwdef mutable struct SizeConfiguration <: Validated
    smallest::Real = 2
    span::Real = 8
end

Configure the range of sizes in pixels (1/96th of an inch) to map the sizes data By default we use 2 pixels for the smallest size, and add to it an additional span of pixel sizes be 8 to get the largest size. Both the smallest size and span of sizes must be positive.

SomeGraphs.Common.LineConfiguration — Type

@kwdef mutable struct LineConfiguration <: Validated
    width::Real = 1
    style::Maybe{LineStyle} = SolidLine
    is_filled::Bool = false
    color::Maybe{AbstractString} = nothing
end

Configure a line in a graph.

If style is nothing no line is drawn. If is_filled then the region defined by the line (below it in a lines graph) is filled. A line with no style and no filled is not drawn but can still be used to define a region (e.g., for a BandsConfiguration ).

By default, the color is chosen automatically.

SomeGraphs.Common.LineStyle — Type

Styles of drawing a line

SolidLine draws a solid line (the default).

DashLine draws a dashed line.

DotLine draws a dotted line.

DashDotLine draws a dash-dotted line.

SomeGraphs.Common.ValuesOrientation — Type

The orientation of the values axis in a distribution(s) or bars graph:

HorizontalValues - The values are the X axis

VerticalValues - The values are the Y axis (the default).

SomeGraphs.Common.BandsConfiguration — Type

@kwdef mutable struct BandsConfiguration <: Validated
    low::BandConfiguration = BandConfiguration(is_dashed = true)
    middle::BandConfiguration = BandConfiguration()
    high::BandConfiguration = BandConfiguration(is_dashed = true)
end

Configure the partition of the graph up to three band regions. The low and high bands are for the "outer" regions (so their lines are at their border, dashed by default) and the middle band is for the "inner" region between them (so its line is inside it, solid by default).

If show_legend , then a legend showing the bands will be shown.

SomeGraphs.Common.BandConfiguration — Type

@kwdef mutable struct BandConfiguration <: Validated
    offset::Maybe{Real} = nothing
    line::LineConfiguration = LineConfiguration()
end

Configure a region of the graph defined by some band of values.

The offset specifies the band's defining line position. We allow up to three bands in a complete BandsConfiguration . A band exists only if its offset is specified, in which case the line specifies how to render its defining line. The low and high bands are defined the region below and above their defining line's offset . If both are defined, the middle band offset defines the center line of the band; the band can still be filled even if this offset is not specified.

SomeGraphs.Common.BandsData — Type

@kwdef mutable struct BandsData
    low_offset::Maybe{Real} = nothing
    middle_offset::Maybe{Real} = nothing
    high_offset::Maybe{Real} = nothing
end

Specify data for bands.

SomeGraphs.Common.ColorsConfiguration — Type

@kwdef mutable struct ColorsConfiguration <: Validated
    show_legend::Bool = false
    color_axis::AxisConfiguration = AxisConfiguration()
    colors_palette::Maybe{Union{AbstractString, ContinuousColors, CategoricalColors}} = nothing
end

Configure how to color some values. The colors_palette is applied; this can be:

The name of a standard Plotly palette (see NAMED_COLOR_PALETTES );

A vector of (value, color) tuples for a continuous (numeric value) scale;

Or a vector of (value, color) for a categorical (string value) scale.

If show_legend is set, then the colors will be shown (in the legend or as a color bar, as appropriate).

The color_axis can be used to tweak (continuous) colors. Specifically, the values to be converted to color are modified according to the log_scale and/or percent flags. This is also done to the values specified in an explicit continuous colors pallete.

In Plotly's API, a color pallete is always specified to cover the values between 0 and 1, and it linearly maps some range of colors to this unit range. You can control the mapping of the values to be colored by using color_axis to modify the values ( log_scale , percent ) and also specify the specific range of color values to map to the unit range ( minimum , maximum ).

When specifying an explict continuous colors pallete, we must map its lowest value to 0 and the highest value to 1 (and the values must be in non-decreasing order). In this case, there's no point in specifying minimum and maximum in the color_axis , as these are already encoded in the pallete itself.

SomeGraphs.Common.ContinuousColors — Type

A continuous colors palette, mapping numeric values to colors. We also allow specifying tuples instead of pairs to make it easy to invoke the API from other languages such as Python which do not have the concept of a Pair . The legend_title is only used if show_legend is set in the configuration.

SomeGraphs.Common.CategoricalColors — Type

A categorical colors palette, mapping string values to colors. An empty string color means the entity will not be shown.

SomeGraphs.Common.NAMED_COLOR_PALETTES — Constant

Builtin color palattes from Plotly , both linear: Blackbody , Bluered , Blues , Cividis , Earth , Electric , Greens , Greys , Hot , Jet , Picnic , Portland , Rainbow , RdBu , Reds , Viridis , YlGnBu , YlOrRd and cyclical: Twilight , IceFire , Edge , Phase , HSV , mrybm , mygbm .

Note

You would think we could just give the builtin color palette names to plotly, but it turns out that "builtin" in Python plotly doesn't mean "builtin" in JavaScript plotly, because "reasons". We therefore have to copy their definition here. An upside of having this dictionary is that you are free to insert additional named palettes into and gain the convenience of refering to them by name (e.g., for coloring heatmap annotations).

A _r suffix specifies reversing the order of the pallete.

You can also append a final _z:<value_fraction>:<color_fraction> suffix to the name. This will map values in the bottom 0.. value_fraction of the range to white, and map the rest of the values to the top color_fraction ..1 range of the palette. For example, Blues_z:0.3:0.2 will color the bottom 30% of the values in white, and color the top 70% of the values to the top 80% of the Blues pallete.

A _c:<value_fraction>:<color_fraction> works similarly to the _z suffix, except that the fractions are centered on 0.5 (the middle) of the values and color ranges. This is meant to be applied when the values range is +/-Diff, and the pallete has white in the middle. For example RdBu_c:0.3:0.2 will color the 30% of the values near the 0 middle in white, and color the rest of the values using the top and bottom 40% of the RdBu pallete.

An _o:<value_fraction>:<color> suffix maps the pallete to the range 0.. 1-value_fraction , and all the values above this to the color , to denote overflow (too high) values. For example, Blues_o:0.01:magenta will color all the values in the bottom 99% of the values range to Blues , and the top 1% of the range to magenta. A _u suffix works similarly but applies to the bottom range of the values.

You can combine multiple suffixes together, for example Reds_z:0.2:0.2_o:0.99:magenta or RdBu_r_c:0.2:0.2_o:0.01:magenta_u:0.01:darkgreen .

Palletes with suffixes (including _r ) are computed on the fly and cached for future use.

Note

The implementation of the suffixes uses 1e-6 as a color difference "too small to matter". Don't use fractions this small in the prefixes or you will have a bad day.

Note

Always hold the COLOR_PALETTES_LOCK when manually accessing the NAMED_COLOR_PALETTES , otherwise you will regret it at some point.

SomeGraphs.Common.COLOR_PALETTES_LOCK — Constant

A global lock to use when accessing the NAMED_COLOR_PALETTES . Always hold the COLOR_PALETTES_LOCK when manually accessing the NAMED_COLOR_PALETTES , otherwise you will regret it at some point.

SomeGraphs.Common.SubGraph — Type

@kwdef struct SubGraph
    index::Integer
    overlay::Bool
end

Identify one sub-graph out of a set of adjacent graphs. If the index is 1, this is the 1st sub-graph (used top initialize some values such as the legend group title). If overlay then the sub-graphs are plotted on top of each other, which affects axis parameters.

Index

SomeGraphs.Common

SomeGraphs.Common.COLOR_PALETTES_LOCK

SomeGraphs.Common.NAMED_COLOR_PALETTES

SomeGraphs.Common.AbstractGraphConfiguration

SomeGraphs.Common.AbstractGraphData

SomeGraphs.Common.AxisConfiguration

SomeGraphs.Common.BandConfiguration

SomeGraphs.Common.BandsConfiguration

SomeGraphs.Common.BandsData

SomeGraphs.Common.CategoricalColors

SomeGraphs.Common.ColorsConfiguration

SomeGraphs.Common.ContinuousColors

SomeGraphs.Common.FigureConfiguration

SomeGraphs.Common.Graph

SomeGraphs.Common.LineConfiguration

SomeGraphs.Common.LineStyle

SomeGraphs.Common.LogScale

SomeGraphs.Common.MarginsConfiguration

SomeGraphs.Common.PlotlyFigure

SomeGraphs.Common.SizeConfiguration

SomeGraphs.Common.SubGraph

SomeGraphs.Common.ValuesOrientation

SomeGraphs.Common.save_graph