Distributions

SomeGraphs.Distributions — Module

Graphs for showing probability distributions.

SomeGraphs.Distributions.DistributionGraph — Type

A graph for visualizing a single distribution. See DistributionGraphData and DistributionGraphConfiguration .

SomeGraphs.Distributions.DistributionConfiguration — Type

@kwdef mutable struct DistributionConfiguration <: Validated
    style::DistributionStyle = CurveDistribution
    show_outliers::Bool = false
    color::Maybe{AbstractString} = nothing
end

Configure the style of a distribution graph.

If style uses a box, then if show_outliers , also show the extreme (outlier) points.

The color is chosen automatically by default. When showing multiple distributions, you can override it per each one in the DistributionsGraphData .

SomeGraphs.Distributions.DistributionStyle — Type

Possible styles for visualizing a distribution:

CurveDistribution - a density curve (the default).

ViolinDistribution - same as a density curve but mirrored below the values axis.

BoxDistribution - a box with whiskers to show important distribution values.

CurveBoxDistribution - combine a curve and a box.

ViolinBoxDistribution - combine a violin and a box.

SomeGraphs.Distributions.DistributionGraphConfiguration — Type

@kwdef mutable struct DistributionGraphConfiguration <: AbstractGraphConfiguration
    figure::FigureConfiguration = FigureConfiguration()
    values_orientation::ValuesOrientation = HorizontalValues
    distribution::DistributionConfiguration = DistributionConfiguration()
    value_axis::AxisConfiguration = AxisConfiguration()
    value_bands::BandsConfiguration = BandsConfiguration()
end

Configure a graph for showing a single distribution.

SomeGraphs.Distributions.DistributionGraphData — Type

@kwdef mutable struct DistributionGraphData <: AbstractGraphData
    figure_title::Maybe{AbstractString} = nothing
    value_axis_title::Maybe{AbstractString} = nothing
    trace_axis_title::Maybe{AbstractString} = nothing
    distribution_values::AbstractVector{<:Real} = Float32[]
    distribution_name::Maybe{AbstractString} = nothing
    distribution_color::Maybe{AbstractString} = nothing
    value_bands::BandsData = BandsData()
end

By default, all the titles are empty. You can specify the overall figure_title as well as the value_axis_title and the trace_axis_title . The optional distribution_name is used as the name of the density axis. You can also specify the distribution_color and/or value_bands offsets here if they are more of a data than a configuration parameter in the specific graph. This will override whatever is specified in the configuration.

SomeGraphs.Distributions.distribution_graph — Function

distribution_graph(;
    [figure_title::Maybe{AbstractString} = nothing,
    value_axis_title::Maybe{AbstractString} = nothing,
    trace_axis_title::Maybe{AbstractString} = nothing,
    distribution_values::AbstractVector{<:Real} = Float32[],
    distribution_name::Maybe{AbstractString} = nothing],
)::DistributionGraph

Create a DistributionGraph by initializing only the DistributionGraphData fields.

SomeGraphs.Distributions.DistributionsGraph — Type

A graph for visualizing multiple distributions. See DistributionsGraphData and DistributionsGraphConfiguration .

SomeGraphs.Distributions.DistributionsGraphConfiguration — Type

@kwdef mutable struct DistributionsGraphConfiguration <: AbstractGraphConfiguration
    figure::FigureConfiguration = FigureConfiguration()
    distribution::DistributionConfiguration = DistributionConfiguration()
    values_orientation::ValuesOrientation = HorizontalValues
    value_axis::AxisConfiguration = AxisConfiguration()
    value_bands::BandsConfiguration = BandsConfiguration()
    distributions_gap::Maybe{Real} = 0.05
    show_legend::Bool = false
end

Configure a graph for showing multiple distributions.

This is similar to DistributionGraphConfiguration , with additions to deal with having multiple distributions.

If distributions_gap is set to nothing , overlay the distributions on top of each other. Otherwise, the distributions are plotted next to each other, with the distributions_gap specified as a fraction of the used graph size. If zero the graphs will be adjacent, if 1 then the gaps will be the same size as the graphs.

If show_legend , we add a legend listing the distributions.

SomeGraphs.Distributions.DistributionsGraphData — Type

@kwdef mutable struct DistributionsGraphData <: AbstractGraphData
    figure_title::Maybe{AbstractString} = nothing
    value_axis_title::Maybe{AbstractString} = nothing
    trace_axis_title::Maybe{AbstractString} = nothing
    legend_title::Maybe{AbstractString} = nothing
    distributions_values::AbstractVector{<:AbstractVector{<:Real}} = Vector{Float32}[]
    distributions_names::Maybe{AbstractVector{<:AbstractString}} = nothing
    distributions_colors::Maybe{AbstractVector{<:AbstractString}} = nothing
end

The data for a multiple distributions graph. By default, all the titles are empty. You can specify the overall figure_title as well as the value_axis_title , the trace_axis_title and the legend_title (if show_legend is set). If specified, the distributions_names and/or the distributions_colors vectors must contain the same number of elements as the number of vectors in the distributions_values .

SomeGraphs.Distributions.distributions_graph — Function

distributions_graph(;
    [figure_title::Maybe{AbstractString} = nothing,
    value_axis_title::Maybe{AbstractString} = nothing,
    trace_axis_title::Maybe{AbstractString} = nothing,
    legend_title::Maybe{AbstractString} = nothing,
    distributions_values::AbstractVector{<:AbstractVector{<:Real}} = Vector{Float32}[],
    distributions_names::Maybe{AbstractVector{<:AbstractString}} = nothing,
    distributions_colors::Maybe{AbstractVector{<:AbstractString}} = nothing],
)::DistributionsGraph

Create a DistributionsGraph by initializing only the DistributionsGraphData fields.

Index

SomeGraphs.Distributions

SomeGraphs.Distributions.DistributionConfiguration

SomeGraphs.Distributions.DistributionGraph

SomeGraphs.Distributions.DistributionGraphConfiguration

SomeGraphs.Distributions.DistributionGraphData

SomeGraphs.Distributions.DistributionStyle

SomeGraphs.Distributions.DistributionsGraph

SomeGraphs.Distributions.DistributionsGraphConfiguration

SomeGraphs.Distributions.DistributionsGraphData

SomeGraphs.Distributions.distribution_graph

SomeGraphs.Distributions.distributions_graph