Heatmaps

SomeGraphs.Heatmaps — Module

Graphs for showing a 2D matrix.

SomeGraphs.Heatmaps.HeatmapGraph — Type

A graph showing a heatmap. See HeatmapGraphData and HeatmapGraphConfiguration .

SomeGraphs.Heatmaps.heatmap_graph — Function

function heatmap_graph(;
    [figure_title::Maybe{AbstractString} = nothing,
    x_axis_title::Maybe{AbstractString} = nothing,
    y_axis_title::Maybe{AbstractString} = nothing,
    entries_colors_title::Maybe{AbstractString} = nothing,
    rows_names::Maybe{AbstractVector{<:AbstractString}} = nothing,
    columns_names::Maybe{AbstractVector{<:AbstractString}} = nothing,
    entries_values::Maybe{AbstractMatrix{<:Real}} = Float32[;;],
    entries_hovers::Maybe{AbstractMatrix{<:AbstractString}} = nothing,
    rows_hovers::Maybe{AbstractVector{<:AbstractString}} = nothing,
    columns_hovers::Maybe{AbstractVector{<:AbstractString}} = nothing,
    rows_annotations::AbstractVector{AnnotationData} = AnnotationData[],
    columns_annotations::AbstractVector{AnnotationData} = AnnotationData[],
    entries_order_by::Maybe{AbstractMatrix{<:Real}} = nothing,
    rows_order_by::Maybe{AbstractMatrix{<:Real}} = nothing,
    columns_order_by::Maybe{AbstractMatrix{<:Real}} = nothing,
    entries_order::Maybe{AbstractVector{<:Integer}} = nothing,
    rows_order::Maybe{AbstractVector{<:Integer}} = nothing,
    columns_order::Maybe{AbstractVector{<:Integer}} = nothing,
    configuration::HeatmapGraphConfiguration = HeatmapGraphConfiguration()]
)::HeatmapGraph

Create a HeatmapGraph by initializing only the HeatmapGraphData fields (with an optional HeatmapGraphConfiguration ).

SomeGraphs.Heatmaps.HeatmapGraphData — Type

@kwdef mutable struct HeatmapGraphData <: AbstractGraphData
    figure_title::Maybe{AbstractString} = nothing
    x_axis_title::Maybe{AbstractString} = nothing
    y_axis_title::Maybe{AbstractString} = nothing
    entries_colors_title::Maybe{AbstractString} = nothing
    rows_names::Maybe{AbstractVector{<:AbstractString}} = nothing
    columns_names::Maybe{AbstractVector{<:AbstractString}} = nothing
    entries_values::Maybe{AbstractMatrix{<:Real}} = Float32[;;]
    entries_hovers::Maybe{AbstractMatrix{<:AbstractString}} = nothing
    rows_hovers::Maybe{AbstractVector{<:AbstractString}} = nothing
    columns_hovers::Maybe{AbstractVector{<:AbstractString}} = nothing
    rows_annotations::AbstractVector{AnnotationData} = AnnotationData[]
    columns_annotations::AbstractVector{AnnotationData} = AnnotationData[]
    entries_order_by::Maybe{AbstractMatrix{<:Real}} = nothing
    rows_order_by::Maybe{AbstractMatrix{<:Real}} = nothing
    columns_order_by::Maybe{AbstractMatrix{<:Real}} = nothing
    entries_order::Maybe{AbstractVector{<:Integer}} = nothing
    rows_order::Maybe{AbstractVector{<:Integer}} = nothing
    columns_order::Maybe{AbstractVector{<:Integer}} = nothing
end

The data for a graph showing a heatmap (matrix) of entries.

This is shown as a 2D image where each matrix entry is a small rectangle with some color. Due to Plotly limitation, colors must be continuous. The hover for each rectangle is a combination of the entries_hovers , rows_hovers and columns_hovers for the entry.

By default, if reordering the data, this is based on the entries_values . You can override this by specifying an entries_order_by matrix of the same size, or even separate rows_order_by and columns_order_by matrices. For efficiency the rows_order_by matrix should be in row-major layout.

Alternatively you can force the order of the data by specifying the entries_order permutation (for a square matrix) or separate rows_order and columns_order permutations.

SomeGraphs.Heatmaps.HeatmapGraphConfiguration — Type

@kwdef mutable struct HeatmapGraphConfiguration <: AbstractGraphConfiguration
    figure::FigureConfiguration = FigureConfiguration()
    value_axis::AxisConfiguration = AxisConfiguration()
    entries_colors::ColorsConfiguration = ColorsConfiguration()
    rows_annotations::AnnotationSize = AnnotationSize()
    columns_annotations::AnnotationSize = AnnotationSize()
    entries_order::Maybe{Reorder} = nothing
    rows_order::Maybe{Reorder} = nothing
    columns_order::Maybe{Reorder} = nothing
    entries_metric::Maybe{PreMetric} = nothing
    rows_metric::Maybe{PreMetric} = nothing
    columns_metric::Maybe{PreMetric} = nothing
end

Configure a graph showing a heatmap.

This displays a matrix of values using a rectangle at each position. Due to Plotly's limitations, you still to manually tweak the graph size for best results; there's no way to directly control the width and height of the rectangles. In addition, the only supported color configurations are using continuous color palettes.

You can use columns_order and/or rows_order to reorder the data. By default, this uses the Euclidean distance metric. You can override this by specifying the columns_metric and/or rows_metric distance measure.

Specifying entries_order is equivalent to specifying both rows_order and columns_order , and similarly for entries_metric .

There's no x_axis or y_axis configuration here, as none of the options apply, except for show_ticks , which is implied by setting the rows_names and/or columns_names of the data. You can however provide default axis titles here (which you can override in the data as usual).

SomeGraphs.Heatmaps.Reorder — Type

Specify how to re-order the rows and/or columns of a matrix for display. Options are:

A symbol which is a valid value for the linkage parameter for hclust (that is, one of :single , :average , :complete , :ward , or :ward_presquared ). TODO: Also allow :slanter .

A symbol which is a valid value for the branchorder parameter for hclust (that is, one of :r , :barjoseph / :optimal ). TODO: Also allow :slanter .

The symbol :same . This can only be used for square matrices, and indicates the order of this axis should be copied from the computed order of the other axis. It is obviously invalid to specify this for both axes, or when the other axis has no order specified.

Examples:

Default (serves as a baseline to compare with when modifying options):

using SomeGraphs
graph = heatmap_graph(; entries_values = [
    0 1 2 3;
    7 6 5 4;
    8 9 10 11;
])
using PlotlyDocumenter
to_documenter(graph.figure)

Annotations:

using SomeGraphs
graph = heatmap_graph(;
    entries_values = [
        0 1 2 3;
        7 6 5 4;
        8 9 10 11;
    ],
    rows_annotations = [
        AnnotationData(;
            title = "is",
            values = ["yes", "maybe", "no"],
            colors = ColorsConfiguration(;
                palette = Dict("yes" => "black", "maybe" => "darkgray", "no" => "lightgray"),
                show_legend = true,
            ),
        ),
    ],
    columns_annotations = [AnnotationData(; title = "score", values = [1, 0.5, 0, 1])],
    entries_colors_title = "values",
    rows_names = ["X", "Y", "Z"],
    columns_names = ["A", "B", "C", "D"],
)
graph.configuration.entries_colors.show_legend = true
using PlotlyDocumenter
to_documenter(graph.figure)

Index

SomeGraphs.Heatmaps

SomeGraphs.Heatmaps.HeatmapGraph

SomeGraphs.Heatmaps.HeatmapGraphConfiguration

SomeGraphs.Heatmaps.HeatmapGraphData

SomeGraphs.Heatmaps.Reorder

SomeGraphs.Heatmaps.heatmap_graph