Utilities

SomeGraphs.Utilities — Module

Utility functions for defining graph types. We do not re-export all symbols from this sub-module to the global MCGraphs namespace.

SomeGraphs.Validations.Maybe — Type

Maybe{T} = Union{T, Nothing}

The type to use when maybe there is a value, maybe there isn't. This is exactly as if writing the explicit Union with Nothing but is shorter and more readable. This is extremely common.

Note

This is replicated from DataAxesFormats to avoid making it a dependency. We do not export it but use it extensively in the type signatures.

SomeGraphs.Common.validate_graph — Function

validate_graph(::Graph)::Maybe{AbstractString}

Validate that the combination of data and configuration in a graph is valid, after validating each one separately. This isn't invoked manually, instead it is called by the overall validate of the graph. It is provided (with a default empty implementation) to allow for type-specific validations.

SomeGraphs.Utilities.graph_layout — Function

graph_layout(figure_configuration::FigureConfiguration, layout::Layout)::Layout

Patch a plotly layout with a figure_configuration .

SomeGraphs.Common.graph_to_figure — Function

graph_to_figure(graph::Graph)::PlotlyFigure

Render a graph given its data and configuration. Technically this just converts the graph to a PlotlyFigure which Julia knows how display for us, rather than actually display the graph. The implementation depends on the specific graph type.

You can just write graph.figure instead of graph_to_figure(graph) .

Note

When saving a figure to a file, Plotly in its infinite wisdom ignores the graph width and height specified inside the figure, (except for saving HTML file). You should therefore use save_graph rather than call savefig on the result of graph_to_figure .

SomeGraphs.Utilities.plotly_figure — Function

plotly_figure(trace::GenericTrace, layout::Layout)::PlotlyFigure
plotly_figure(traces::AbstractVector{<:GenericTrace}, layout::Layout)::PlotlyFigure

Wrap a trace or a set of traces with the accompanying layout in a PlotlyFigure .

SomeGraphs.Utilities.axis_ticksformat — Function

axis_ticksformat(axis_configuration::AxisConfiguration)::Tuple{Maybe{AbstractString}, Maybe{AbstractString}}

Given the axis_configuration , return a prefix and a suffix for the ticks along that axis. This deals with log scales and percent scaling.

SomeGraphs.Utilities.scale_axis_value — Function

scale_axis_value(axis_configuration::AxisConfiguration, value::Maybe{Real})::Maybe{Real}

Scale a single value according to the axis_configuration . This deals with log scales and percent scaling.

SomeGraphs.Utilities.scale_axis_values — Function

scale_axis_values(
    axis_configuration::AxisConfiguration,
    values::Union{AbstractVector{<:Real},AbstractVector{<:Maybe{Real}}},
)::Union{AbstractVector{<:Real},AbstractVector{<:Maybe{Real}}}

Scale a vector of values according to the axis_configuration . This deals with log scales and percent scaling.

SomeGraphs.Utilities.horizontal_bands_shapes — Function

horizontal_bands_shapes(
    axis_configuration::AxisConfiguration,
    scaled_values_range::AbstractVector{<:Real},
    bands_data::BandsData,
    bands_configuration::BandsConfiguration
)::AbstractVector{<:Shape}

Return a vector of shapes for plotting horizontal bands. These shapes need to be places in the layout and not the traces because Plotly.

SomeGraphs.Utilities.vertical_bands_shapes — Function

vertical_bands_shapes(
    axis_configuration::AxisConfiguration,
    scaled_values_range::AbstractVector{<:Real},
    bands_data::BandsData,
    bands_configuration::BandsConfiguration
)::AbstractVector{<:Shape}

Return a vector of shapes for plotting vertical bands. These shapes need to be places in the layout and not the traces because Plotly.

Index

SomeGraphs.Utilities

SomeGraphs.Validations.Maybe

SomeGraphs.Common.graph_to_figure

SomeGraphs.Common.validate_graph

SomeGraphs.Utilities.axis_ticksformat

SomeGraphs.Utilities.graph_layout

SomeGraphs.Utilities.horizontal_bands_shapes

SomeGraphs.Utilities.plotly_figure

SomeGraphs.Utilities.scale_axis_value

SomeGraphs.Utilities.scale_axis_values

SomeGraphs.Utilities.vertical_bands_shapes