Utilities

SomeGraphs.Utilities — Module

Utility functions for defining graph types. We do not re-export all symbols from this sub-module to the global MCGraphs namespace. You can safely ignore these unless you are implementing a new graph type.

SomeGraphs.Validations.Maybe — Type

Maybe{T} = Union{T, Nothing}

The type to use when maybe there is a value, maybe there isn't. This is exactly as if writing the explicit Union with Nothing but is shorter and more readable. This is extremely common.

Note

This is replicated from DataAxesFormats to avoid making it a dependency. We do not export it but use it extensively in the type signatures.

SomeGraphs.Common.validate_graph — Function

validate_graph(::Graph)::Maybe{AbstractString}

Validate that the combination of data and configuration in a graph is valid, after validating each one separately. This isn't invoked manually, instead it is called by the overall validate of the graph. It is provided (with a default empty implementation) to allow for type-specific validations.

SomeGraphs.Common.graph_to_figure — Function

graph_to_figure(graph::Graph)::PlotlyFigure

Render a graph given its data and configuration. Technically this just converts the graph to a PlotlyFigure which Julia knows how display for us, rather than actually display the graph. The implementation depends on the specific graph type.

You can just write graph.figure instead of graph_to_figure(graph) .

Note

When saving a figure to a file, Plotly in its infinite wisdom ignores the graph width and height specified inside the figure, (except for saving HTML file). You should therefore use save_graph rather than call savefig on the result of graph_to_figure .

SomeGraphs.Utilities.plotly_figure — Function

plotly_figure(trace::GenericTrace, layout::Layout)::PlotlyFigure
plotly_figure(traces::AbstractVector{<:GenericTrace}, layout::Layout)::PlotlyFigure

Wrap a trace or a set of traces with the accompanying layout in a PlotlyFigure .

SomeGraphs.Utilities.validate_colors — Function

validate_colors(
    colors_data_context::ValidationContext,
    colors_data::Maybe{Union{AbstractVector{<:AbstractString}, AbstractVector{<:Real}}},
    colors_configuration_context::ValidationContext,
    colors_configuration::ColorsConfiguration,
    mask::Maybe{Union{AbstractVector{Bool},BitVector}} = nothing,
)::Nothing

Validate that the colors_data from the colors_data_context is valid and consistent with the colors_configuration from the colors_configuration_context . For example, if the color configuration contains a categorical color mapping, this will validate that all the color names in the data are valid keys of this mapping.

If a mask is specified, do not validate colors in the data whose matching value in the mask is false.

SomeGraphs.Utilities.fill_color — Function

fill_color(line_color::Maybe{AbstractString})::Maybe{AbstractString}

Return a fill color based on a line_color . The fill color is twice as transparent as the line color.

SomeGraphs.Utilities.validate_values — Function

validate_values(
    values_data_context::ValidationContext,
    values_data::Maybe{AbstractVector{<:Real}}},
    axis_configuration_context::ValidationContext,
    axis_configuration::AxisConfiguration,
)::Nothing

Validate that the values_data from the values_data_context is valid and consistent with the axis_configuration from the axis_configuration_context . Specifically this ensures that if a log_scale is applied, all the values are positive.

SomeGraphs.Utilities.scale_axis_value — Function

scale_axis_value(axis_configuration::AxisConfiguration, value::Real; clamp::Bool = true)::Real
scale_axis_value(axis_configuration::AxisConfiguration, value::Nothing; clamp::Bool = true)::Nothing

Scale a single value according to the axis_configuration . This deals with log scales and percent scaling. By default, clamp the values to a specified explicit range.

SomeGraphs.Utilities.scale_axis_values — Function

scale_axis_values(
    axis_configuration::AxisConfiguration,
    values::Maybe{AbstractVector{<:Maybe{Real}}};
    clamp::Bool = true
)::Maybe{AbstractVector{<:Maybe{Real}}}

Scale a vector of values according to the axis_configuration . This deals with log scales and percent scaling By default, clamp the values to a specified explicit range.

SomeGraphs.Utilities.scale_size_values — Function

scale_size_values(
    sizes_configuration::SizesConfiguration,
    values::Maybe{AbstractVector{<:Real}},
)::Maybe{AbstractVector{<:Real}}

Scale a vector of values according to sizes_configuration .

SomeGraphs.Utilities.plotly_layout — Function

plotly_layout(
    figure_configuration::FigureConfiguration;
    title::Maybe{AbstractString},
    showlegend::Bool,
    shapes::AbstractVector{Shape},
)::Layout

Create a Plotly Layout object.

SomeGraphs.Utilities.set_layout_axis! — Function

set_layout_axis!(
    layout::Layout,
    axis::AbstractString
    axis_configuration::AxisConfiguration;
    title::Maybe{AbstractString},
    range::Range,
)::Nothing

Add a Plotly axis in a layout using the axis_configuration .

SomeGraphs.Utilities.set_layout_colorscale! — Function

set_layout_colorscale!(
    layout::Layout,
    colorscale::AbstractString,
    colors_configuration::ColorsConfiguration;
    offset::Maybe{Real},
    title::Maybe{AbstractString},
)::Nothing

Set a colorscale in a Plotly layout , as specified by a colors_configuration . Since Plotly is dumb when it comes to placement of color scales, the offset must be specified manually to avoid overlaps.

SomeGraphs.Utilities.plotly_line_dash — Function

plotly_line_dash(line_style::LineStyle)::Maybe{AbstractString}

Return the Plotly line_dash for a line_style .

SomeGraphs.Utilities.validate_graph_bands — Function

validate_graph_bands(
    field::AbstractString,
    bands_configuration::BandsConfiguration,
    bands_data::BandsData,
    axis_configuration::AxisConfiguration,
)::Nothing

Validate that the bands configuration and data is compatible. Assumes these are specified as the same field in both the graph's data and configuration.

SomeGraphs.Utilities.push_horizontal_bands_shapes — Function

push_horizontal_bands_shapes(
    shapes::AbstractVector{Shape},
    axis_configuration::AxisConfiguration,
    scaled_values_range::Range,
    bands_data::BandsData,
    bands_configuration::BandsConfiguration
)::AbstractVector{<:Shape}

Push shapes for plotting horizontal bands. These shapes need to be placed in the layout and not the traces because Plotly.

SomeGraphs.Utilities.push_vertical_bands_shapes — Function

push_vertical_bands_shapes(
    shapes::AbstractVector{Shape},
    axis_configuration::AxisConfiguration,
    scaled_values_range::Range,
    bands_data::BandsData,
    bands_configuration::BandsConfiguration
)::AbstractVector{<:Shape}

Push shapes for plotting vertical bands. These shapes need to be places in the layout and not the traces because Plotly.

SomeGraphs.Utilities.push_diagonal_bands_shapes — Function

push_diagonal_bands_shapes(
    shapes::AbstractVector{Shape},
    x_axis_configuration::AxisConfiguration,
    y_axis_configuration::AxisConfiguration,
    x_scaled_values_range::Range,
    y_scaled_values_range::Range,
    bands_data::BandsData,
    bands_configuration::BandsConfiguration
)::AbstractVector{<:Shape}

Push shapes for plotting diagonal bands. These shapes need to be placed in the layout and not the traces because Plotly.

SomeGraphs.Utilities.MaybeRange — Type

@kwdef mutable struct MaybeRange
    minimum::Maybe{Float32} = nothing
    maximum::Maybe{Float32} = nothing
end

A range of values (possibly partially specified).

SomeGraphs.Utilities.Range — Type

@kwdef mutable struct Range
    minimum::Float32
    maximum::Float32
end

A range of values (fully specified).

SomeGraphs.Utilities.collect_range! — Function

collect_range!(
    range::MaybeRange,
    values::AbstractVector{<:Maybe{Real}},
)::Nothing

Expand the range to cover the `values.

SomeGraphs.Utilities.expand_range! — Function

expand_range!(range::Maybe{AbstractVector{<:Maybe{Real}}})::Nothing

Expand the range of values by 1% to allow for points and lines at the edge to be fully visible.

SomeGraphs.Utilities.assert_range — Function

assert_range(range::MaybeRange)::Range

Convert a partial range to a fully specified one.

SomeGraphs.Utilities.final_scaled_range — Function

final_scaled_range(
    implicit_scaled_range::Range,
    axis_configuration::AxisConfiguration
)::Range,

Compute the final range for some axis given the implicit_scaled_range computed from the values and the axis_configuration .

SomeGraphs.Utilities.prefer_data — Function

prefer_data(data_value::Any, configuration_value::Any)::Any
prefer_data(data_values::AbstractVector, index::Integer, configuration_value::Any)::Any

Return a value to use, prefering the data value (which may be in a vector) to the configuration value.

Index

SomeGraphs.Utilities

SomeGraphs.Utilities.MaybeRange

SomeGraphs.Utilities.Range

SomeGraphs.Validations.Maybe

SomeGraphs.Common.graph_to_figure

SomeGraphs.Common.validate_graph

SomeGraphs.Utilities.assert_range

SomeGraphs.Utilities.collect_range!

SomeGraphs.Utilities.expand_range!

SomeGraphs.Utilities.fill_color

SomeGraphs.Utilities.final_scaled_range

SomeGraphs.Utilities.plotly_figure

SomeGraphs.Utilities.plotly_layout

SomeGraphs.Utilities.plotly_line_dash

SomeGraphs.Utilities.prefer_data

SomeGraphs.Utilities.push_diagonal_bands_shapes

SomeGraphs.Utilities.push_horizontal_bands_shapes

SomeGraphs.Utilities.push_vertical_bands_shapes

SomeGraphs.Utilities.scale_axis_value

SomeGraphs.Utilities.scale_axis_values

SomeGraphs.Utilities.scale_size_values

SomeGraphs.Utilities.set_layout_axis!

SomeGraphs.Utilities.set_layout_colorscale!

SomeGraphs.Utilities.validate_colors

SomeGraphs.Utilities.validate_graph_bands

SomeGraphs.Utilities.validate_values