Utilities

SomeGraphs.Utilities — Module

Utility functions for defining graph types. We do not re-export all symbols from this sub-module to the global MCGraphs namespace. You can safely ignore these unless you are implementing a new graph type.

SomeGraphs.Validations.Maybe — Type

Maybe{T} = Union{T, Nothing}

The type to use when maybe there is a value, maybe there isn't. This is exactly as if writing the explicit Union with Nothing but is shorter and more readable. This is extremely common.

Note

This is replicated from DataAxesFormats to avoid making it a dependency. We do not export it but use it extensively in the type signatures.

SomeGraphs.Common.validate_graph — Function

validate_graph(::Graph)::Maybe{AbstractString}

Validate that the combination of data and configuration in a graph is valid, after validating each one separately. This isn't invoked manually, instead it is called by the overall validate of the graph. It is provided (with a default empty implementation) to allow for type-specific validations.

SomeGraphs.Utilities.final_scaled_range — Function

final_scaled_range(
    implicit_scaled_range::AbstractVector{<:Maybe{Real}},
    axis_configuration::AxisConfiguration
)::AbstractVector{<:Real}

Compute the final range for some axis given the implicit_scaled_range computed from the values and the axis_configuration .

SomeGraphs.Utilities.patch_layout_figure! — Function

patch_layout_figure!(layout::Layout, figure_configuration::FigureConfiguration)::Nothing

Patch a Plotly layout using the figure_configuration .

SomeGraphs.Utilities.patch_layout_axis! — Function

patch_layout_axis!(
    layout::Layout,
    axis::AbstractString
    axis_configuration::AxisConfiguration,
)::Nothing

Patch an existing Plotly axis in a layout using the axis_configuration .

SomeGraphs.Common.graph_to_figure — Function

graph_to_figure(graph::Graph)::PlotlyFigure

Render a graph given its data and configuration. Technically this just converts the graph to a PlotlyFigure which Julia knows how display for us, rather than actually display the graph. The implementation depends on the specific graph type.

You can just write graph.figure instead of graph_to_figure(graph) .

Note

When saving a figure to a file, Plotly in its infinite wisdom ignores the graph width and height specified inside the figure, (except for saving HTML file). You should therefore use save_graph rather than call savefig on the result of graph_to_figure .

SomeGraphs.Utilities.plotly_figure — Function

plotly_figure(trace::GenericTrace, layout::Layout)::PlotlyFigure
plotly_figure(traces::AbstractVector{<:GenericTrace}, layout::Layout)::PlotlyFigure

Wrap a trace or a set of traces with the accompanying layout in a PlotlyFigure .

SomeGraphs.Utilities.axis_ticks_prefix — Function

axis_ticks_prefix(axis_configuration::AxisConfiguration)::Maybe{AbstractString}

Given the axis_configuration , return a prefix for the ticks along that axis. This deals with log scales.

SomeGraphs.Utilities.axis_ticks_suffix — Function

axis_ticks_suffix(axis_configuration::AxisConfiguration)::Maybe{AbstractString}

Given the axis_configuration , return a suffix for the ticks along that axis. This deals with percent scaling.

SomeGraphs.Utilities.scale_axis_value — Function

scale_axis_value(axis_configuration::AxisConfiguration, value::Real; clamp::Bool = true)::Real
scale_axis_value(axis_configuration::AxisConfiguration, value::Nothing; clamp::Bool = true)::Nothing

Scale a single value according to the axis_configuration . This deals with log scales and percent scaling. By default, clamp the values to a specified explicit range.

SomeGraphs.Utilities.scale_axis_values — Function

scale_axis_values(
    axis_configuration::AxisConfiguration,
    values::Maybe{AbstractVector{<:Maybe{Real}}};
    clamp::Bool = true
)::Maybe{AbstractVector{<:Maybe{Real}}}

Scale a vector of values according to the axis_configuration . This deals with log scales and percent scaling By default, clamp the values to a specified explicit range.

SomeGraphs.Utilities.scale_size_values — Function

scale_size_values(
    axis_configuration::AxisConfiguration,
    size_configuration::SizeConfiguration,
    values::Maybe{AbstractVector{<:Real}},
)::Maybe{AbstractVector{<:Real}}

Scale a vector of values according to the axis_configuration and size_configuration .

SomeGraphs.Utilities.push_horizontal_bands_shapes — Function

push_horizontal_bands_shapes(
    shapes::AbstractVector{Shape},
    axis_configuration::AxisConfiguration,
    scaled_values_range::AbstractVector{<:Real},
    bands_data::BandsData,
    bands_configuration::BandsConfiguration
)::AbstractVector{<:Shape}

Push shapes for plotting horizontal bands. These shapes need to be placed in the layout and not the traces because Plotly.

SomeGraphs.Utilities.push_vertical_bands_shapes — Function

push_vertical_bands_shapes(
    shapes::AbstractVector{Shape},
    axis_configuration::AxisConfiguration,
    scaled_values_range::AbstractVector{<:Real},
    bands_data::BandsData,
    bands_configuration::BandsConfiguration
)::AbstractVector{<:Shape}

Push shapes for plotting vertical bands. These shapes need to be places in the layout and not the traces because Plotly.

SomeGraphs.Utilities.push_diagonal_bands_shapes — Function

push_diagonal_bands_shapes(
    shapes::AbstractVector{Shape},
    x_axis_configuration::AxisConfiguration,
    y_axis_configuration::AxisConfiguration,
    x_scaled_values_range::AbstractVector{<:Real},
    y_scaled_values_range::AbstractVector{<:Real},
    bands_data::BandsData,
    bands_configuration::BandsConfiguration
)::AbstractVector{<:Shape}

Push shapes for plotting diagonal bands. These shapes need to be placed in the layout and not the traces because Plotly.

SomeGraphs.Utilities.validate_graph_bands — Function

validate_graph_bands(
    field::AbstractString,
    bands_configuration::BandsConfiguration,
    bands_data::BandsData,
    axis_configuration::AxisConfiguration,
)::Nothing

Validate that the bands configuration and data is compatible. Assumes these are specified as the same field in both the graph's data and configuration.

SomeGraphs.Utilities.validate_colors — Function

validate_colors(
    colors_data_context::ValidationContext,
    colors_data::Maybe{Union{AbstractVector{<:AbstractString}, AbstractVector{<:Real}}},
    colors_configuration_context::ValidationContext,
    colors_configuration::ColorsConfiguration,
    mask::Maybe{Union{AbstractVector{Bool},BitVector}} = nothing,
)::Maybe{AbstractString}

Validate that the colors_data from the colors_data_context is valid and consistent with the colors_configuration from the colors_configuration_context . For example, if the color configuration contains a categorical color mapping, this will validate that all the color names in the data are valid keys of this mapping.

If a mask is specified, do not validate colors in the data whose matching value in the mask is false.

SomeGraphs.Utilities.fill_color — Function

fill_color(line_color::Maybe{AbstractString})::Maybe{AbstractString}

Return a fill color based on a line_color . The fill color is twice as transparent as the line color.

SomeGraphs.Utilities.range_of — Function

range_of(
    values::AbstractVector{<:Maybe{Real}},
    range::Maybe{AbstractVector{<:Maybe{Real}}} = nothing,
)::AbstractVector{<:Maybe{Real}}

Compute the range of (scaled) values. If a range is give, it is updated in-place as well.

SomeGraphs.Utilities.expand_range! — Function

expand_range!(range::Maybe{AbstractVector{<:Maybe{Real}}})::Nothing

Expand the range of values by 1% to allow for points and lines at the edge to be fully visible.

SomeGraphs.Utilities.prefer_data — Function

prefer_data(data_value::Any, configuration_value::Any)::Any
prefer_data(data_values::AbstractVector, index::Integer, configuration_value::Any)::Any

Return a value to use, prefering the data value (which may be in a vector) to the configuration value.

Index

SomeGraphs.Utilities

SomeGraphs.Validations.Maybe

SomeGraphs.Common.graph_to_figure

SomeGraphs.Common.validate_graph

SomeGraphs.Utilities.axis_ticks_prefix

SomeGraphs.Utilities.axis_ticks_suffix

SomeGraphs.Utilities.expand_range!

SomeGraphs.Utilities.fill_color

SomeGraphs.Utilities.final_scaled_range

SomeGraphs.Utilities.patch_layout_axis!

SomeGraphs.Utilities.patch_layout_figure!

SomeGraphs.Utilities.plotly_figure

SomeGraphs.Utilities.prefer_data

SomeGraphs.Utilities.push_diagonal_bands_shapes

SomeGraphs.Utilities.push_horizontal_bands_shapes

SomeGraphs.Utilities.push_vertical_bands_shapes

SomeGraphs.Utilities.range_of

SomeGraphs.Utilities.scale_axis_value

SomeGraphs.Utilities.scale_axis_values

SomeGraphs.Utilities.scale_size_values

SomeGraphs.Utilities.validate_colors

SomeGraphs.Utilities.validate_graph_bands